Remove `Region` from `mir::Rvalue::Ref`
This was only used for borrow checking, but we will create a fresh
`RegionVid` for each borrow anyway, so there's no need for us to store
it in the MIR.

Also removes `Region`s from various places where they're no longer
neede. Uses `ReErased` for any regions that need to be created in MIR
generation.